+++ commons/dimas-core/src/lib.rs
@@ -2,10 +2,6 @@
 
 //! Core of `DiMAS`
 
-/// Communication
-pub mod communicator;
-/// Configuration
-pub mod config;
 /// Error handling
 pub mod error;
 /// Utilities